Break commandline parsing out of main.cpp
